Fix 'This backend doesn't support absolute paths.' prod error
This error would occur in production every time when cleaning the
metadata of an uploaded file. The reason for this is that the Django
storage backend `S3Storage` that we're using in production doesn't
support accessing the `path` attribute of files. Files in S3 don't have
filepaths, but instead the name of the file is the pseudo "path" to the
file.