Fix 'Could not get a libmat2 parser for filetype image/jpeg.' prod error
The error occurred whenever uploading files in production.
It happened because the file name that we were passing to
`get_parser` was an S3 path which didn't actually exist on the disk.
This fixes it by doing the metadata cleaning in a temporary file.

Also changed to simplify the metadata cleaning by doing it before the
instance is saved. This was not possible before because libmat2 requires
the target file to exists on disk.