Generate API docs from resolver docstrings and class attributes
This makes documenting query resolvers a lot more convent since one
doesn't need to use the clumsy `description` argument, but instead the
documentation can exists straight in the docstring, where it really
belongs.

This also makes it so that some general features of resolvers and
mutations, such as weather they require logging in or verification to
use, can be taken from the code's properties straight to the docs
without redundantly explicitly including that information also in the
`description` argument, which could always introduce errors and mismatch
between the code and the docs.

Basically nothing should change in the existing API docs from a reader's
perspective.