Remove AUTOCOMPLETE_MAX_RESULTS slicing from queries with no filtering
Since the frontend cannot filter the output of these queries, it could
just result in a situation where we have for example 55 cities, but the
last 5 of them (alphabetically) are just completely inaccessible by the
frontend.

We don't solve this by adding the `name` parameter to these resolvers,
since that would make the frontend requery these constantly
unnecessarily when the user types more characters to the field. The
current implementation for these queries in the frontend is super
efficient, since it just makes one request for the field, and then does
rest of the filtering on the client.