Remove the "implicit" `order_by("pk")` from SkoleManager
It's more efficient and clearer to just sort the querysets when its
needed.

Also remove the resourceTypes query that would now after 5ab8b5a be
identical to the autocompleteResourceTypes query. Frontend was also not
using this at all.

Do not order the queries in MyData for added performance, it most likely
does not matter at all if those results are not consistently ordered.
This can always be done later if deemed necessary.