Allow querying autocomplete courses by the code as well
In the frontend we show the autocomplete results as:
`Test Engineering Course 18 - TEST00018`, so it's only logical that the
code part of the results can also be used for the filtering.

Maybe we should make this change to the course schema as well, so that
one could search by course code also from the frontpage course search
bar. Maybe then we could remove the code filtering completely from the
search page as well?