Make `0005_add_slug_fields` data migration populate the slug correctly
Previously Schools would get slugs like `'school-object-1'`. The reason
for this was that migrations do not have access to the model methods, so
the `populate_from="__str__"` of `AutoSlugField` didn't work as
expected.

> Because it’s impossible to serialize arbitrary Python code, these
> historical models will not have any custom methods that you have defined.
>
> https://docs.djangoproject.com/en/stable/topics/migrations/#historical-models

Another challenge was that we needed to temporarily bypass the
`always_update=True` of `AutoSlugField`, because it would just otherwise
overwrite the slug with the (incorrect) autogenerated value.

Also moved the `forwards_func` into its own migration, it's best
practice to separate schema and data migrations, and this allows us to
get this upgraded function ran in the QA automatically, since 0005 was
already executed there.

This still has the risk that if a future migration saves the slug field
of some instances, the slugs of them will then again get overwritten
with the autogenerated junk.