Only allow registering with certain email address domains
The current set of `ALLOWED_EMAIL_DOMAINS` are the student emails of all
Finnish universities.

For convenience also change all test email domains to be `@test.test`,
so that we have to whitelist only that.

This also fixes a bug where if a user tried to update the casing of
their username or email they would get the error that the username/email
is already taken.