Require verifying user's backup email with the primary one
Otherwise an attacker who was looking to change the password of a user,
with access to its account settings, could simply change the backup
email to be their (=the attackers) own one and then just request a
password reset email to it. This would completely bypass the safeguard
that to change the password one needs to either know the old password or
have access to the trusted primary email.