Fix incorrectly resolved conflicts and make InviteCode a one2one
We already had the reverse relationship named as `invite_code` and not
`invite_codes`, and the business logic states that we only need one, so
let's make it a proper one to one. Otherwise we would annoyingly need to
use `user.invite_code.first()` in some places, and even if we would
rename the relation to `invite_codes` using `first()` on it would still
be ugly.