Improve quoting in Dockerfile's `poetry install` step
The previous quoting was logically wrong; we simply closed the quotes
before the subshell expansion and then had an extra pair of quotes at
the end with nothing in them.

It would've been logical to use:
```
\"$([ "$install_dev_dependencies" -eq 0 ] && printf -- '--no-dev')\"
```
to prevent word globbing but unfortunately it leads to this error when
`$install_dev_dependencies` is set to 1:
```
$ poetry install --no-root ""
Too many arguments.
```